cmd/snap-update-ns: move new entry methods here
